[BE] release: v1.2.2 배포

+++ backend/src/main/java/feedzupzup/backend/organization/dto/response/CheeringResponse.java
@@ -5,10 +5,10 @@ import io.swagger.v3.oas.annotations.media.Schema;
 @Schema(description = "응원하기 응답")
 public record CheeringResponse(
         @Schema(description = "단체의 응원 총 개수", example = "200")
-        int cheeringTotalCount
+        long cheeringTotalCount
 ) {
 
-        public static CheeringResponse from(final int cheeringTotalCount) {
+        public static CheeringResponse from(final long cheeringTotalCount) {
                 return new CheeringResponse(cheeringTotalCount);
         }
 }
